+++ src/screens/validator_details/components/transactions/types.ts
@@ -1,0 +1,6 @@
+export type TransactionState = {
+  hasNextPage: boolean;
+  isNextPageLoading: boolean;
+  offsetCount: number;
+  data: Transactions[];
+}
